refactor(lotto): 로또 같은 숫자 비교 결과 자료형 int -> long 변경

+++ src/main/java/domain/Lotto.java
@@ -37,8 +37,8 @@ public class Lotto {
                 .count();
     }
 
-    public int getSameNumberCount(Lotto anotherLotto) {
-        return (int) lottoNumbers.stream()
+    public long getSameNumberCount(Lotto anotherLotto) {
+        return lottoNumbers.stream()
                 .filter(anotherLotto::containsLottoNumber)
                 .count();
     }
